// api/src/ai/tools/schema.ts
export const PrimaryKeyInputSchema = z.union([z.number(), z.string()]);
export const PrimaryKeyValidateSchema = z.union([z.number(), z.string()]);
// ...
export const TriggerFlowInputSchema = z.object({
    id: PrimaryKeyInputSchema,
    collection: z.string(),
    keys: z.array(PrimaryKeyInputSchema).optional(),
    headers: z.record(z.string(), z.any()).optional(),
    query: z.record(z.string(), z.any()).optional(),
    data: z.record(z.string(), z.any()).optional(),
});

export const TriggerFlowValidateSchema = z.strictObject({
    id: PrimaryKeyValidateSchema,
    collection: z.string(),
    keys: z.array(PrimaryKeyValidateSchema).optional(),
    query: z.record(z.string(), z.any()).optional(),
    headers: z.record(z.string(), z.any()).optional(),
    data: z.record(z.string(), z.any()).optional(),
});